Move BundleValidate to a standalone method and call from DBOLM
In order for DeployableByOLM to check that the bundle is valid before
actually deploying, this moves the bundle validate code to a shared
area, so it can be called by both the DeployableByOLM check and the
actual Bundle Validate check.

DeployableByOLM does not do extensive checking of the result of
BundleValidate. If there is an error, or Passed: false is in the
report, it will fail.
